+++ src/app/components/vio/form/VioFormContact.vue
@@ -1,5 +1,10 @@
 <template>
-  <VioForm :form="v$" :is-form-sent="isFormSent" @submit="submit">
+  <VioForm
+    :form="v$"
+    :is-loading="isLoading"
+    :is-form-sent="isFormSent"
+    @submit.prevent="submit"
+  >
     <input type="hidden" name="static-form-name" value="contact" />
     <VioFormInput
       id-label="input-name"
@@ -61,6 +66,10 @@
         </VioFormInputStateError>
       </template>
     </VioFormInput>
+    <VioFormInputCaptcha
+      :form-input="v$.captcha"
+      @input="form.captcha = $event"
+    />
   </VioForm>
 </template>
 
@@ -68,12 +77,28 @@
 import { useVuelidate } from '@vuelidate/core'
 import { required } from '@vuelidate/validators'
 
+type FormValid = { emailAddress: string; name: string; message: string }
+
+withDefaults(
+  defineProps<{
+    isLoading?: boolean
+  }>(),
+  {
+    isLoading: undefined,
+  },
+)
+
+const emit = defineEmits<{
+  submit: [form: FormValid]
+}>()
+
 const { t } = useI18n()
 const runtimeConfig = useRuntimeConfig()
 const siteConfig = useSiteConfig()
 
 // data
 const form = reactive({
+  captcha: ref<string>(),
   emailAddress: ref<string>(),
   name: ref<string>(),
   message: ref<string>(),
@@ -81,11 +106,16 @@ const form = reactive({
 const isFormSent = ref(false)
 
 // methods
-const submit = async (e: Event) =>
-  !(await isFormValid({ v$, isFormSent })) ? e.preventDefault() : undefined
+const submit = async () => {
+  if (!(await isFormValid({ v$, isFormSent }))) return
+  emit('submit', form as FormValid)
+}
 
 // vuelidate
 const rules = {
+  captcha: {
+    required,
+  },
   emailAddress: {
     // maxLength: maxLength(100),
     required,