+++ src/server/utils/constants.ts
@@ -1,1 +1,0 @@
-export { IS_IN_FRONTEND_DEVELOPMENT } from '../../node'